+++ investment-calculator/src/App.jsx
@@ -4,11 +4,27 @@ import UserInput from './components/UserInput';
 import ResultsTable from './components/ResultsTable';
 
 function App() {
+	const [userInput, setUserInput] = useState({
+		initialInvestment: 10000,
+		annualInvestment: 1200,
+		expectedReturn: 4,
+		duration: 6,
+	});
+
+	const handleInputChange = (label, event) => {
+		setUserInput((prev) => {
+			return {
+				...prev,
+				[label]: event.target.value,
+			};
+		});
+	};
+
 	return (
 		<>
 			<Header />
 			<main>
-				<UserInput />
+				<UserInput inputValues={userInput} onInputChange={handleInputChange} />
 				<ResultsTable />
 			</main>
 		</>
